Fix compilation in Linux

+++ include/morphsnakes/ndimage.h
@@ -5,6 +5,7 @@
 #include <iostream>
 #include <array>
 #include <numeric>
+#include <algorithm>
 
 namespace morphsnakes
 {
@@ -19,7 +20,7 @@ template<size_t D>
 std::array<int, D> operator+(const std::array<int, D>& a, const std::array<int, D>& b)
 {
     std::array<int, D> res;
-    for(int i = 0; i < D; ++i)
+    for(size_t i = 0; i < D; ++i)
         res[i] = a[i] + b[i];
     return res;
 }
@@ -59,7 +60,7 @@ std::ostream& operator<<(std::ostream& ostr, const morphsnakes::Position<2>& pos
 template<size_t D>
 bool isBoundary(const std::array<int, D>& coords, const Shape<D>& shape)
 {
-    for(int i = 0; i < D; ++i)
+    for(size_t i = 0; i < D; ++i)
     {
         if(coords[i] == 0 || coords[i] == shape[i]-1)
             return true;
